Refactor `buildGenericTxHistory` in `txBridgeM`
This refactor allows to reuse the expected generalized object for different
bridges in multiples functions and different files.
